refactor(actual): own variation transaction shapes in a neutral module
ExistingVariationTransaction, VariationTransactionInput, and the variation
constants move from the reconciliation policy into a neutral module so the
Actual adapter can produce and the policy can consume the same clean shapes
without either importing the other.
