+++ src/actual/reconciliation-policy.ts
@@ -1,26 +1,13 @@
-export const VARIATION_NOTES = "Variation"
-export const VARIATION_IMPORTED_ID_PREFIX = "fintual-variation:"
+import {
+  VARIATION_IMPORTED_ID_PREFIX,
+  VARIATION_NOTES,
+  type ExistingVariationTransaction,
+  type VariationTransactionInput,
+} from "./variation-transaction.ts"
 
 export interface BalanceEntry {
   date: number
   real_difference: number
-}
-
-export interface ExistingVariationTransaction {
-  id: string
-  date?: string
-  notes?: string
-  payee?: string | null
-  imported_id?: string
-}
-
-export interface VariationTransactionInput {
-  date: string
-  amount: number
-  payee?: string
-  notes: string
-  imported_id: string
-  cleared: boolean
 }
 
 export type ReconciliationAction =
